Rework on how DC handles dev, dev:online, test; Vite moved to a Docker container
