+++ _Scrap/app-deploy-ops/.images/gcp-api-restrictions.png
