+++ _Scrap/app-deploy-ops/sh/gen-firebase-ci.sh
