+++ _Scrap/app-deploy-ops/src/ops-adapters/cloudLogging/index.js
