+++ _Scrap/app-deploy-ops/src/ops-adapters/cloudLogging/worker.js
