+++ ci/cloudbuild.backend.yaml
@@ -82,7 +82,7 @@ steps:
 # --- Timeout
 #
 # Cloud Build:
-#   - 2:17, 2:15
+#   - 2:17, 2:15, 2:46
 #
 timeout: 300s
