+++ packages/app/.gitignore
@@ -1,9 +1,8 @@
 # 'node_modules' ignored by the main project
 
+# tbd. dim out /dist/
 /dist/
-
-# Firebase hosting caches
-/.firebase/
+/public/dist/
 
 # (follow project root's .gitignore on those)
 #/.firebaserc
