+++ packages/app/dev/index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
-    <link rel="icon" type="image/png" href="/favicon.png">
+    <link rel="icon" type="image/png" href="../public/favicon.png">
     <!--
     <link rel="icon" href="data:;base64,iVBORw0KGgo=">  <!_-- no favicon --_>
     -->
@@ -17,7 +17,7 @@
     </style>
 
     <script type="module">
-      import '/init/main-dev.js';
+      import '/src/main.js';
     </script>
   </head>
