+++ packages/app/package.json
@@ -20,9 +20,12 @@
   "scripts": {
     "dev": "npm run dev:local",
     "-- ": "",
-    "test": "npm run -s _emulRunning && npm run -s _genEnvLocal && (docker compose run -e SENTRY_DNS= --rm vite-local-launched) && CYPRESS_defaultCommandTimeout=10000 cypress run",
+    "test": "npm run -s _emulRunning && npm run -s _testDC && npm run -s _testRun",
+    "_testDC": "docker compose -f docker-compose.test.yml run --rm vite-for-test-launched",
+    "_testRun": "CYPRESS_defaultCommandTimeout=10000 cypress run",
+    "dc:launch:test": "npm run -s _genEnvLocal && vite --mode dev_local",
     "-- -": "",
-    "dev:local": "npm run -s _checkNoEsbuild && npm run -s _emulPrimed && docker compose down vite-local && docker compose run --rm --service-ports vite-local",
+    "dev:local": "npm run -s _checkNoEsbuild && npm run -s _emulPrimed && docker compose -f docker-compose.yml -f docker-compose.local.yml run --rm --service-ports vite-dev",
     "_emulRunning": "(cd ../backend && npm run app:start)",
     "_emulPrimed": "npm run -s _emulRunning && npm run -s _prime",
     "_prime": "docker compose run --rm prime",
@@ -31,20 +34,20 @@
     "--  -": "",
     "ci:launch": "vite --mode dev_local",
     "-- --": "",
-    "dev:online": "docker compose -f docker-compose.online.yml run --rm --service-ports vite-3001",
+    "dev:online": "npm run -s _checkStaging && npm run -s _checkNoEsbuild && docker compose -f docker-compose.yml -f docker-compose.online.yml run --rm --service-ports vite-dev",
     "dc:launch:online": "npm run -s _genEnvOnline && vite --mode dev_online",
     "_genEnvOnline": "gen-vite-env-online > .env",
     "--   -": "",
     "lint": "eslint --ext .js,.vue src || true",
     "-- - -": "",
-    "build": "vite build",
+    "build": "vite build -c vite.config.js",
+    "serve": "docker compose run --rm --service-ports _serve",
     "--  --": "",
-    "//_checkStaging": "[[ -f ../../firebase.${ENV-staging}.js ]] || ( >&2 echo \"ERROR: '../../firebase.${ENV-staging}.js' is missing; see 'README.md' on how to create one.\n\"; false)",
+    "_checkStaging": "[[ -f ../../firebase.${ENV-staging}.js ]] || ( >&2 echo \"ERROR: '../../firebase.${ENV-staging}.js' is missing; see 'README.md' on how to create one.\n\"; false)",
     "_checkNoEsbuild": "[[ ! -d node_modules/esbuild ]] || ( >&2 echo \"INTERNAL ERROR: 'node_modules/esbuild*' shouldn't be there; please remove them. (rm -rf node_modules/esbuild*)\"; false) ",
     "-- ---": "",
-    "postbuild": "cp public/* dist/",
     "postinstall": "if [ -z $BUILDER_OUTPUT ]; then npm run -s _postinstall_dev ; else npm run -s _postinstall_ci ; fi",
-    "_postinstall_dev": "rm -rf node_modules/esbuild node_modules/esbuild-* && npm install --prefix .. --no-save esbuild",
+    "_postinstall_dev": "rm -rf node_modules/vite node_modules/esbuild node_modules/esbuild-*",
     "_postinstall_ci": "npm run -s _genEnvCi",
     "//_postinstall_ci": "npm install --prefix .. --no-save esbuild && npm run -s _genEnvCi)",
     "_genEnvCi": "FIREBASE_JSON=../backend/firebase.app.json EMUL_HOST=emul-for-app gen-vite-env-local --project=demo-main > .env"
@@ -54,7 +57,7 @@
     "@sentry/tracing": "^6.19.7",
     "aside-keys": "^0.0.3-alpha.1",
     "firebase": "^9.8.1",
-    "vue": "^3.2.34",
+    "vue": "^3.2.35",
     "vue-router": "^4.0.15"
   },
   "//dependencies": {
@@ -67,7 +70,6 @@
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-vue": "^9.0.1",
     "prettier": "2.6.2",
-    "vite": "^2.9.9",
     "vue-eslint-parser": "^9.0.2"
   },
   "//comments": {
@@ -116,10 +118,7 @@
         "command here, mostly for just documenting the use."
       ],
       "_prime": [
-        "The project name used _must_ match that in '../backend/', used for launching the emulators.",
-        "",
-        "To develop the 'prime.dc', rebuild by:",
-        " 'docker compose build prime && docker compose run --rm prime'"
+        "The project name used _must_ match that in '../backend/', used for launching the emulators."
       ],
       "_touchEm": [
         "Vital for DC that files to be mapped to the container exist *before* the DC command."
