+++ packages/app/public/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+    <link rel="icon" type="image/png" href="/favicon.png">
+    <!--
+    <link rel="icon" href="data:;base64,iVBORw0KGgo=">  <!_-- no favicon --_>
+    -->
+
+    <script type="module">
+      import '/dist/app.js';
+    </script>
+  </head>
+
+  <body>
+    <div id="app"></div>
+  </body>
+</html>
