+++ packages/app/tools/serve.dc/README.md
@@ -1,0 +1,6 @@
+# `serve.dc`
+
+Minimal local hosting of files built for deployment.
+
+This can be used as a manual step to check the output, before needing to deploy.
+
