+++ packages/app/tools/vite.dc/README.md
@@ -1,0 +1,3 @@
+# `vite.dc`
+
+Development serving (with Hot Module Reload) and production build of a web app.
