--- packages/app/vitebox/public
@@ -1,1 +1,0 @@
-../public
